1.5 moved enums to a different directory
