+++ conveyor/src/main/java/com/neoflex/conveyor/dto/enumType/EmailThemeType.java
@@ -1,4 +1,4 @@
-package com.neoflex.conveyor.dto;
+package com.neoflex.conveyor.dto.enumType;
 
 public enum EmailThemeType {
     WELCOME,
