+++ conveyor/src/main/java/com/neoflex/conveyor/dto/enumType/EmploymentStatusType.java
@@ -1,12 +1,9 @@
-package com.neoflex.conveyor.dto;
+package com.neoflex.conveyor.dto.enumType;
 
 public enum EmploymentStatusType {
     EMPLOYED,
     UNEMPLOYED,
     BUSINESS_OWNER,
-    JUNIOR,
-    MIDDLE,
-    SENIOR,
     RETIRED,
     OTHER
 }
