+++ conveyor/src/main/java/com/neoflex/conveyor/dto/enumType/PositionType.java
@@ -1,11 +1,13 @@
-package com.neoflex.conveyor.dto;
+package com.neoflex.conveyor.dto.enumType;
 
 public enum PositionType {
     MANAGER,
     ENGINEER,
     ANALYST,
     ASSISTANT,
+    JUNIOR,
+    MIDDLE,
+    SENIOR,
     DIRECTOR,
-    SPECIALIST,
     OTHER
 }
